requirements.txt
Flask
python-dotenv
simple-salesforce
pandas
plotly
gunicorn